refactor(scan): fix cycles in the scan service
